requirements.txt
asgiref==3.3.4
autopep8==1.5.5
certifi==2020.12.5
Django==3.2
numpy==1.20.1
olefile==0.46
Pillow==8.2.0
pip==21.0.1
psycopg2==2.8.6
pycodestyle==2.6.0
pytz==2021.1
setuptools==52.0.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
torch==1.7.1+cpu
torchvision==0.8.2+cpu
typing-extensions==3.7.4.3
wheel==0.36.2
wincertstore==0.2
